Rename epub_controller.dart to epub_controller_base.dart
